requirements.txt
jupyter
numpy
pandas
matplotlib
yfinance
tqdm
scikit-learn
networkx
python-louvain
openpyxl
duckdb
pyarrow
python-dotenv
streamlit
pytest
openai
